% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani_distance.R
\name{ani_distance}
\alias{ani_distance}
\title{ani_distance}
\usage{
ani_distance(
  se,
  taxonomy = NULL,
  tax_level = "Species",
  collapse_method = "representative"
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object.}

\item{taxonomy}{An optional \code{data.frame} object containing the taxonomy information for the samples in \code{se}.
The result of running read_taxonomy on the taxonomy file associated with the ANI data.
If provided, the ANI values will be collapsed at the specified taxonomic level before calculating the distance matrix. (default=NULL)}

\item{tax_level}{A character string specifying the taxonomic level at which to collapse the ANI values. (default="Species")}

\item{collapse_method}{A character string specifying the method to use to collapse the ANI values. Must be one of 'representative', 'mean' or 'max'. (default='representative')}
}
\value{
A distance matrix.
}
\description{
This function calculates the distance between samples in a \code{SummarizedExperiment} object representing strain level ANI data.
}
\examples{
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
tax_path <- system.file("extdata", "example_taxonomy.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path)
tax <- read_taxonomy(tax_path)
distance_matrix <- ani_distance(se)

}
