% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\docType{class}
\name{betaGLM-class}
\alias{betaGLM-class}
\title{betaGLM: A Class for Generalized Linear Models with Zero-Inflation}
\description{
The \code{betaGLM} class is designed to store the results of strainspy
}
\section{Slots}{

\describe{
\item{\code{row_data}}{A \code{DataFrame} containing row data.}

\item{\code{coefficients}}{A \code{DataFrame} containing model coefficients.}

\item{\code{std_errors}}{A \code{DataFrame} containing standard errors of the coefficients.}

\item{\code{p_values}}{A \code{DataFrame} containing p-values for the coefficients.}

\item{\code{zi_coefficients}}{A \code{DataFrame} containing zero-inflation coefficients.}

\item{\code{zi_std_errors}}{A \code{DataFrame} containing zero-inflation standard errors.}

\item{\code{zi_p_values}}{A \code{DataFrame} containing zero-inflation p-values.}

\item{\code{fdr}}{A \code{DataFrame} or \code{NULL}, containing false discovery rates (FDR).}

\item{\code{zi_fdr}}{A \code{DataFrame} or \code{NULL}, containing zero-inflation FDR values.}

\item{\code{dispersion}}{A numeric vector containing dispersion estimates for the model.}

\item{\code{residuals}}{A \code{DataFrame} containing residuals from the model.}

\item{\code{design}}{A matrix representing the design matrix of the model.}

\item{\code{assay}}{A matrix containing the assay data used for fitting the model.}

\item{\code{call}}{The matched call of the model.}
}}

