% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_EffectSizes.R
\name{comp_ani_diff_and_posthoc_test}
\alias{comp_ani_diff_and_posthoc_test}
\title{Compute the mean ANI differences for contigs with an identity effect and
perform posthoc testing to establish support for each hit.}
\usage{
comp_ani_diff_and_posthoc_test(
  se,
  fit,
  th = NULL,
  simplified_formula = NULL,
  beta_min_nz = 0.1,
  beta_min_ani_diff = 0.015,
  nthreads = 1L,
  reorder_hits = FALSE,
  BPPARAM = NULL
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing
assay data and metadata}

\item{fit}{A fitted \code{strainspy_fit}model}

\item{th}{\code{tibble} returned by \code{top_hits()}.}

\item{simplified_formula}{A formula to use instead of \code{fit@design}, useful
when the original design was too complex and causes emmeans to fail during
ANI estimation (default NULL).}

\item{beta_min_nz}{Minimum non-zero proportion per phenotype group required
to consider a beta hit is well supported. Hits driven by a proportion
of non zero values below this threshold at least in one phenotype group are
flagged \verb{too few nonzero}. Default 0.1. Hits driven by less than or equal to
3 observations will be flagged automatically \verb{<3 nonzero}.}

\item{beta_min_ani_diff}{Minimum absolute mean ANI difference required  for
a beta hit to be considered meaningful. Effect sizes small than this are
flagged: \verb{small effect}. Default = 1.5e-3 corresponds to 0.15\% delta ANI.}

\item{nthreads}{Number of threads for parallel processing. Default \code{1}.}

\item{reorder_hits}{Logical. Return \code{th} after reordering hits to priorities
best supported ones. Default \code{FALSE}.}

\item{BPPARAM}{Optional BiocParallelParam object.}
}
\value{
A \code{tibble} similar to \code{th}, with added columns giving details on:
Component (Beta, ZI or both), Mean ANI_Difference, Contrast, Non-Zero Counts
of contrast components, and a general \code{Comment} on each hit. The commetn will
be \code{NA} if the hit passes all posthoc tests.
}
\description{
In the zero-inflated beta model, beta effects represent changes in
strain identity and are modeled on the logit scale. While this scale
is effective for distinguishing phenotype groups statistically, the
resulting effect sizes are not directly interpretable as differences
in mean ANI. Same applies for effect sizes from the ordinal beta model.
}
\details{
This function converts logit-scale beta effects into response-scale
mean ANI differences using marginal means and pairwise contrasts
(requires \code{emmeans}). It therefore provides an interpretable estimate of the
magnitude of strain-level differences between phenotype groups, along
with post hoc quality checks for poorly supported beta signals.

Posthoc testing uses two approaches. First, it counts the number of non-zero
observations supporting each beta effect. Second, it compares the estimated
mean ANI difference against a preset threshold. A comment is added to each
hit depending on the outcome.

By default, this function uses the formula from \code{fit@design}. However, if the
formula contains multiple covariates, this can cause emmeans to behave
unexpectedly. For example, if \code{fit@design = ~ disease + age + sex + country + (1|study)},
this can cause this function to throw errors or warnings.  Using a \code{simplified_formula = ~ disease}
to override can be a useful approach.

This function is designed to directly work on the \code{tibble} returned from
\code{top_hits()}.
}
\examples{
if (requireNamespace("emmeans", quietly = TRUE)) {
  meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
  package = "strainspy"))
  meta$Case_status <- factor(meta$Case_status)
  se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
  package = "strainspy"), meta_data = meta)
  se <- filter_by_presence(se, min_nonzero = 30)[1:10, ]
  fit <- glmZiBFit(se, ~ Case_status + Age_at_collection, nthreads = 1)
  th <- top_hits(fit, coef = 2, alpha = 1)
  posthoc <- comp_ani_diff_and_posthoc_test(se, fit, th, nthreads = 1)
  head(posthoc$ANI_Difference)
}

}
