% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_priors.R
\name{define_priors}
\alias{define_priors}
\title{Constructor for strainspy_priors using preset or user defined priors}
\usage{
define_priors(
  se,
  design,
  method = "preset_weak",
  add_dispersion_prior = FALSE,
  priors_df = NULL
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object with count matrix and colData.}

\item{design}{A fixed-effects-only formula (e.g., ~ group + age + sex).}

\item{method}{Prior type ("preset_weak", "preset_strong", "manual"). To compute
empirical priors, use \code{compute_eb_priors()}. Default \code{preset_weak}.}

\item{add_dispersion_prior}{Add a prior of normal(0,1) to the dispersion parameter. Default False}

\item{priors_df}{NULL if \code{method == 'preset_weak' | 'preset_strong'}, else a
data.frame of priors suitable for glmmTMB model fitting. See \code{?glmmTMB::priors}
for more details. Default NULL.}
}
\value{
An object of class \code{strainspy_priors}.
}
\description{
Constructor for strainspy_priors using preset or user defined priors
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz",
 package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
pri <- suppressWarnings(define_priors(se[1:10, ], 
~ Case_status + Age_at_collection, method = "preset_weak"))
class(pri)

}
