% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eb_MAP_priors.R
\name{ebayesian_priors}
\alias{ebayesian_priors}
\title{Constructor for strainspy_priors using empirical Bayes estimation}
\usage{
ebayesian_priors(sy, design, chunk_size = 500)
}
\arguments{
\item{sy}{A \code{SummarizedExperiment} object with count matrix and colData.}

\item{design}{A fixed-effects-only formula (e.g., ~ group + age + sex).}

\item{chunk_size}{Integer. Number of strains to process per chunk. Default: 500.}
}
\value{
An object of class \code{strainspy_priors}.
}
\description{
Fits fast binomial and beta models to estimate MAP prior SDs for use in
zero-inflated and ordinal beta models fitted using \code{GLMMTMB}.
}
