% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_EffectSizes.R
\name{estimate_effect_sizes}
\alias{estimate_effect_sizes}
\title{Estimate mean ANI differences for selected contigs analysed using a Zero
Inflated Beta model.}
\usage{
estimate_effect_sizes(
  se,
  ZB_fit,
  th = NULL,
  phenotype_of_interest = NULL,
  top_hit_contigs = NULL,
  beta_min_nz = 0.1,
  beta_min_ani_diff = 0.01,
  nthreads = 1L,
  scale_continuous = TRUE,
  BPPARAM = NULL
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing
assay data and metadata}

\item{ZB_fit}{A fitted \code{strainspy_fit} zero-inflated beta model}

\item{th}{Optional \code{tibble} returned by \code{top_hits()}. If this is provided,
\code{top_hit_contigs} and \code{phenotype_of_interest} are extracted automatically.}

\item{phenotype_of_interest}{Character string specifying the phenotype.
Required if \code{th = NULL}. Default \code{NULL}}

\item{top_hit_contigs}{Character vector of contig names to analyse.
Required if \code{th = NULL}. Default \code{NULL}}

\item{beta_min_nz}{Minimum non-zero proportion per phenotype group required
to consider a beta hit is well supported. Hits driven by a proportion
of non zero values below this threshold at least in one phenotype group are
flagged \verb{too few nonzero}. Default 0.1.}

\item{beta_min_ani_diff}{Minimum absolute mean ANI difference required  for
a beta hit to be considered meaningful. Effect sizes small than this are
flagged: \verb{small effect}. Default 0.01.}

\item{nthreads}{Number of threads for parallel processing. Default \code{1}.}

\item{scale_continuous}{Logical, whether to scale numeric columns in colData.
Default \code{TRUE}}

\item{BPPARAM}{Optional BiocParallelParam object.}
}
\value{
A data.frame with columns:
\code{Contig_name}, \code{Contrast}, \code{Min_NonZero_Ratio}, \code{ANI_Difference},
\code{Significant_Component}, and \code{Comment}.
}
\description{
In the zero-inflated beta model, beta effects represent changes in
strain identity and are modeled on the logit scale. While this scale
is effective for distinguishing phenotype groups statistically, the
resulting effect sizes are not directly interpretable as differences
in mean ANI.
}
\details{
This function converts logit-scale beta effects into response-scale
mean ANI differences using marginal means and pairwise contrasts
(requires \code{emmeans}). It therefore provides an interpretable estimate of the
magnitude of strain-level differences between phenotype groups, along
with post hoc quality checks for poorly supported beta signals.

Contigs can be supplied directly via \code{top_hit_contigs}, or indirectly
by passing the output of \code{top_hits()} via \code{th}. If \code{th} is provided,
\code{phenotype_of_interest} and \code{top_hit_contigs} must be \code{NULL}.
}
