% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_strainspy_priors-class.R
\name{extract_fixef_priors}
\alias{extract_fixef_priors}
\alias{extract_fixef_priors,strainspy_priors-method}
\title{Access fixef priors needed for ordinal beta regression}
\usage{
extract_fixef_priors(object)

\S4method{extract_fixef_priors}{strainspy_priors}(object)
}
\arguments{
\item{object}{A \code{strainspy_priors} object.}
}
\value{
A data frame of prior parameters for fixed effects.
}
\description{
Access fixef priors needed for ordinal beta regression
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
pri <- suppressWarnings(define_priors(se[1:10, ], 
~ Case_status + Age_at_collection, method = "preset_weak"))
head(extract_fixef_priors(pri))
}
