% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_presence.R
\name{filter_by_presence_longitudinal}
\alias{filter_by_presence_longitudinal}
\title{Filter strains based on subject and time aware manner - useful for longitudinal datasets}
\usage{
filter_by_presence_longitudinal(
  se,
  min_subjects = 3,
  min_timepoints = 2,
  subject_col = "subject",
  time_col = "time"
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object.}

\item{min_subjects}{Integer. Minimum number of subjects that must have the strain.}

\item{min_timepoints}{Integer. Minimum number of non-zero timepoints per subject.Default is 3.}

\item{subject_col}{Character. Column name in \code{colData(se)} corresponding to subject IDs. Default is 2.}

\item{time_col}{Character. Column name in \code{colData(se)} corresponding to timepoints.}
}
\value{
A filtered \code{SummarizedExperiment} object with only the rows that meet the criteria.
}
\description{
FThis function filters a SummarizedExperiment object to keep only rows (features/strains)
present in at least \code{min_subjects} individuals and at least \code{min_timepoints} timepoints
per subject.
}
