% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceFit.R
\name{fit_lm}
\alias{fit_lm}
\title{Fit a linear model for a Single Feature}
\usage{
fit_lm(se_subset, col_data, combined_formula)
}
\arguments{
\item{se_subset}{A \code{SummarizedExperiment} object containing the assay data.}

\item{col_data}{A data frame containing the design matrix and additional covariates.}

\item{combined_formula}{The formula for the conditional mean model.}
}
\value{
A list with model coefficients, zero-inflation coefficients, residuals,
log-likelihood, and convergence status.
}
\description{
Fits a simple linear model for functions without random effects using stats::lm()
}
