% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colors}
\alias{get_colors}
\title{Generate colours for Manhattan plots}
\usage{
get_colors(n)
}
\arguments{
\item{n}{Integer. The number of colours needed.}
}
\value{
A character vector of \code{n} colour hex codes.
}
\description{
This function generates a vector of colours for plotting, prioritising a fixed palette of 21 distinct colours.
If more than 21 colours are required, additional colours are generated using \code{grDevices::rainbow()}.
}
