% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{get_confusion_mx}
\alias{get_confusion_mx}
\title{Generate a confusion matrix for a fitted model}
\usage{
get_confusion_mx(top_hits, gt_contigs, all_contigs, print_cm = TRUE)
}
\arguments{
\item{top_hits}{Output from \code{tophits(fit)}}

\item{gt_contigs}{Vector of ground truth contigs}

\item{all_contigs}{Vector of all contigs}

\item{print_cm}{logical. Print the confusion matrix (default = TRUE)}
}
\value{
Confusion Matrix.
}
\description{
Generate a confusion matrix for a fitted model
}
