% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_priors-class.R
\name{getPriors}
\alias{getPriors}
\alias{getPriors,strainspy_priors-method}
\title{Access prior dataframe}
\usage{
getPriors(object)

\S4method{getPriors}{strainspy_priors}(object)
}
\arguments{
\item{object}{strainspy_priors object}
}
\value{
Data frame of prior parameters
}
\description{
Access prior dataframe
}
