% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_strainspy_fit-class.R
\name{getResiduals}
\alias{getResiduals}
\alias{getResiduals,strainspy_fit-method}
\title{Access Residuals}
\usage{
getResiduals(object)

\S4method{getResiduals}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A \code{DataFrame} containing residuals.
}
\description{
Returns the residuals from a \code{strainspy_fit} object.
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
fit <- glmZiBFit(filter_by_presence(se, min_nonzero = 30)[1:10, ], 
~ Case_status + Age_at_collection, nthreads = 1, return_vcov = FALSE)
class(getResiduals(fit))
}
