% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_strainspy_fit-class.R
\name{getSamples}
\alias{getSamples}
\alias{getSamples,strainspy_fit-method}
\title{Get Sample names in a fit object}
\usage{
getSamples(object)

\S4method{getSamples}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A vector of fitted sample names
}
\description{
Gets the ordered list of samples used in \code{strainspy_fit}
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
fit <- glmZiBFit(filter_by_presence(se, min_nonzero = 30)[1:10, ], 
~ Case_status + Age_at_collection, nthreads = 1, return_vcov = FALSE)
head(getSamples(fit))
}
