% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_strainspy_fit-class.R
\name{getZICoefficients}
\alias{getZICoefficients}
\alias{getZICoefficients,strainspy_fit-method}
\title{Access Zero-Inflation Coefficients}
\usage{
getZICoefficients(object)

\S4method{getZICoefficients}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A \code{DataFrame} containing zero-inflation coefficients.
}
\description{
Returns the zero-inflation coefficients from a \code{strainspy_fit} object.
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
fit <- glmZiBFit(filter_by_presence(se, min_nonzero = 30)[1:10, ], 
~ Case_status + Age_at_collection, nthreads = 1, return_vcov = FALSE)
class(getZICoefficients(fit))
}
