% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmObFit.R
\name{glmObFit}
\alias{glmObFit}
\title{glmObFit}
\usage{
glmObFit(
  se,
  design,
  nthreads = 1L,
  scale_continuous = TRUE,
  MAP_prior = "preset_weak",
  family = glmmTMB::ordbeta(),
  BPPARAM = NULL
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{design}{Formula. A formula to specify the fixed and random effects, e.g., \code{ ~ Group + (1|Sample)}.}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{scale_continuous}{Logical. If \code{TRUE}, all numeric columns in \code{colData(se)}
are z-score standardized (mean = 0, SD = 1). Defaults to \code{FALSE}.}

\item{MAP_prior}{One of \code{strainspy_priors} object, character string (\verb{"preset_weak", "preset_strong"}),
a \code{data.frame} of priors, or \code{NULL}. Default \code{"preset_weak"}. See Details.}

\item{family}{A \code{glmmTMB} family object. Defaults to \code{glmmTMB::ordbeta()}.
For general use, leave this at default Advanced users can experiment with others families in \code{?glmmTMB::family_glmmTMB}}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}
}
\value{
A \code{strainspy_fit} object with the following components:
\item{row_data}{A DFrame with 6 slots with feature details}
\item{coefficients}{A DFrame with coefficients for each feature}
\item{std_errors}{A DFrame of standard errors for each feature}
\item{p_values}{A DFrame of p-values for each feature}
\item{residuals}{A DFrame of residual vectors for each feature}
\item{convergence}{A named logical vector indicating convergence for each feature}
\item{design}{Formula used in the call to \code{glmObFit}}
\item{call}{Call to \code{glmObFit}}
}
\description{
TODO: Implement return of vcov info. Patched for compatibility (returns NULL)
}
\details{
This function fits a ordinalbeta regression model using the \code{glmmTMB} package.
It takes a \code{SummarizedExperiment} object as input, along with a user-defined formula,
and fits an ordinal beta regression model on the assay data.

For robust model fitting, it is recommended to use maximum a posteriori (MAP)
estimation. Emprical Bayes MAP priors can be computed using \code{strainspy::compute_eb_priors()}.
They can also be manually defined using \code{strainspy::define_priors()}. The resulting
\code{strainspy_priors} object can be passed to the \code{MAP_prior} argument.

Alternatively, a character string (\code{"preset_weak"} or \code{"preset_strong"}) or a
valid prior \code{data.frame} (see \code{?glmmTMB::priors}) can be provided instead.
This will result in a call to \code{strainspy::define_priors()} to generate priors.
To disable MAP estimation and perform standard REML fitting, set \code{MAP_prior = NULL}.
}
\examples{
\donttest{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)

design <- as.formula(" ~ Case_status + Age_at_collection")

fit <- glmObFit(se[1:10,],  design, family=glmmTMB::ordbeta())

}

}
