% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadjust.R
\name{hadjust}
\alias{hadjust}
\title{Performs hierarchical taxonomy aware adjustment of p-values.}
\usage{
hadjust(object, coef = 2, method = "HMP", taxonomy = NULL, index_range = FALSE)
}
\arguments{
\item{object}{A \code{strainspy_fit} object.}

\item{coef}{The coefficient to use for p-value adjustment. Defaults to 2.}

\item{method}{The method for p-value adjustment. Options include "bonferroni" and Harmonic Mean P-value (HMP).}

\item{taxonomy}{A taxonomy data.table object. If provided, the p-values will be adjusted at each taxonomic level.}

\item{index_range}{This option is only used to generate Manhattan plot segments for higher taxonomic levels}
}
\value{
A tibble with original and adjusted p-values.
}
\description{
This function performs multiple testing adjustment on the results of a betaGLM object
integrating taxonomic information. Various correction methods are available,
including Harmonic Mean P-value (HMP), Bonferroni and Benjamini-Hochberg (BH).
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
se <- filter_by_presence(se, min_nonzero = 30)
tax <- read_taxonomy(system.file("extdata", "example_taxonomy.tsv.gz",
package = "strainspy"))
fit <- glmZiBFit(se[1:10, ], ~ Case_status + Age_at_collection, 
nthreads = 1, return_vcov = FALSE)
out <- hadjust(fit, coef = 2, taxonomy = tax)
head(out)

}
