% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metaphlan.R
\name{merge_metaphlan_files}
\alias{merge_metaphlan_files}
\title{merge_metaphlan_files}
\usage{
merge_metaphlan_files(
  metaphlan_files,
  output_folder,
  output_filename = "metaphlan",
  sgb_database = TRUE
)
}
\arguments{
\item{metaphlan_files}{A character vector of file paths to merge. Optionally, if the input is named, \code{names(metaphlan_files)} will be used as sample identifiers in place of the filename.}

\item{output_folder}{Folder path to save outputs.}

\item{output_filename}{File name for the two outputs. By default, the two files will be saved as metaphlan_merged.tsv.gz and metaphlan_taxonomy.tsv.gz}

\item{sgb_database}{bool. Specify whether the default sgb database was used for profiling. Set to FALSE if \code{sgb_to_gtdb_profile.py} was run on the profiles as a post processing step. Default TRUE}
}
\value{
Invisibly returns \code{NULL}. The function writes two files to \code{output_folder}:
a merged profile table (\verb{*_merged.tsv.gz}) and a taxonomy table (\verb{*_taxonomy.tsv.gz}).
}
\description{
Merge \code{MetaPhlAn >=4.0} profiles into a single file to use with \code{read_metaphlan()} and generate the taxonomy file required for downstream analysis.
}
\examples{
files <- character(0)
is.character(files)

}
