% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{offset_ANI}
\alias{offset_ANI}
\title{Offset ANI by a fixed value}
\usage{
offset_ANI(x, eps = 0.01)
}
\arguments{
\item{x}{Vector of ANI values in (0,1)}

\item{eps}{offset value (default 1e-2)}
}
\value{
Adjusted ANI vector
}
\description{
This function reduces all non-zero ANI values by \code{eps}
}
