% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ani_distribution.R
\name{plot_ani_dist}
\alias{plot_ani_dist}
\title{Generate a violin or box plot for a subset of contigs for a categorical phenotype.}
\usage{
plot_ani_dist(
  se,
  phenotype,
  contigs,
  facet_phenotype = NULL,
  contig_names = NULL,
  show_zero_plot = TRUE,
  show_points = TRUE,
  plot = TRUE,
  plot_type = "box"
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{phenotype}{Any factor variable in \code{colnames(se@colData)}.}

\item{contigs}{A vector of contigs to generate distribution plots}

\item{facet_phenotype}{Any factor variable in \code{colnames(se@colData)}, except \code{phenotype}. If provided, this will be used to facet the plot. See \code{?ggplot2::facet_wrap()}. Default NULL.}

\item{contig_names}{Optional character vector to replace long contig names in the plot.
Must match the length and order of \code{contigs}. If NULL (default), shortening will be attempted automatically.}

\item{show_zero_plot}{Bool. If TRUE, distribution of the proportion of zeros will also be plotted. Default TRUE}

\item{show_points}{Bool. If TRUE, points will be overlaid with jitter using ggforce::geom_sina(). Default TRUE}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot.}

\item{plot_type}{Specify whether \code{box} or \code{violin} plot is required. Default \code{box}.}
}
\value{
If \code{plot = TRUE}, returns a \code{ggplot} object (or a combined patchwork plot
when \code{show_zero_plot = TRUE}). If \code{plot = FALSE}, returns a long-format table
used to generate the ANI distribution plot.
}
\description{
Generate a violin or box plot for a subset of contigs for a categorical phenotype.
}
\examples{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- read.csv(example_meta_path)
# Change variable from character to factor
example_meta$Case_status = factor(example_meta$Case_status)

example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)

plot_ani_dist(se, phenotype = "Case_status", contigs = rownames(se)[1:5])

}
