% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ani_distribution.R
\name{plot_ani_distribution}
\alias{plot_ani_distribution}
\title{plot_ani_distribution}
\usage{
plot_ani_distribution(se, genome_file, predictor, plot = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{genome_file}{Character. The name of the genome file (contig) to extract from the dataset.}

\item{predictor}{Character. The phenotype to group samples in the plot.}

\item{plot}{Logical. If \code{TRUE} (default), the function returns a ggplot object.
If \code{FALSE}, it returns a data frame containing the plotted data.}
}
\value{
A ggplot object displaying a boxplot of ANI values with a beeswarm overlay across predictor categories.
}
\description{
Generate a Boxplot for ANI Values of a Given Contig and Phenotype
}
\details{
This function generates a boxplot to visualize the ANI (Average Nucleotide Identity) values
for a specific contig across different levels of a given categorical predictor.
}
\examples{
\dontrun{
example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
plot_ani_distribution(se, genome_file = 'GCF_025147485.1', predictor = 'Case_status')
}

}
