% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ani_distribution.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Generate a histogram for a single contig for a categorical phenotype.}
\usage{
plot_histogram(
  se,
  phenotype,
  contig,
  bins = 250,
  drop_zeros = FALSE,
  show_quantiles = TRUE,
  fit_spline = FALSE,
  fit_control = 100,
  separate_facets = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{phenotype}{One variable in \code{colnames(se@colData)}, except \code{Sample_File}. The selected variable must be categorical}

\item{contig}{Contig name to visualise}

\item{bins}{Numeric. Number of bins to plot the histogram. This parameter should be adjusted to properly visualise the ANI differences. Default 250}

\item{drop_zeros}{Bool. If TRUE, ANI or Abundance 0 values will not be included in the violin plot. Default FALSE}

\item{show_quantiles}{Bool. If TRUE, min, max, etc. will be shown as vertical lines. Default TRUE}

\item{fit_spline}{Bool. If TRUE, a spline will be fitted to each histogram. Default FALSE}

\item{fit_control}{Numeric. Smaller values increase the smoothness of the fitted spline. Default 100}

\item{separate_facets}{Bool. If TRUE, each category of the phenotype will be plotted in a separate facet. Default TRUE}
}
\value{
A \code{ggplot} histogram object. A summary table of per-group statistics is
attached as \code{attr(plot, "summary_table")}.
}
\description{
Generate a histogram for a single contig for a categorical phenotype.
}
\examples{
\donttest{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
# Change variable from character to factor
example_meta$Case_status = factor(example_meta$Case_status)

example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path, example_meta)

plot_histogram(se, phenotype = "Case_status", contig = rownames(se)[1])

}

}
