% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Generate a strain Manhattan plot from a strainspy_fit object}
\usage{
plot_manhattan(
  object,
  coef = 2,
  taxonomy = NULL,
  aggregate_by_taxa = NULL,
  method = "holm",
  alpha = 0.05,
  tax_levels = c("Phylum", "Genus", "Species", "Strain"),
  plot = TRUE
)
}
\arguments{
\item{object}{A \code{strainspy_fit} object.}

\item{coef}{The number of the coefficient from which to generate the plot (default=2).}

\item{taxonomy}{An optional taxonomy file read using strainspy::read_taxonomy (default=NULL).}

\item{aggregate_by_taxa}{Logical. If TRUE, aggregate p-values across \code{tax_levels} and visualise as a stacked segment plot. Requires \code{taxonomy}. If NULL (default), the function automatically sets it to TRUE when \code{taxonomy} is provided.}

\item{method}{Character. Multiple testing correction method for p-values (e.g., "holm"). Defaults to "holm". Only applicable if \code{aggregate_by_taxa = TRUE}.}

\item{alpha}{Numeric. Significance threshold for adjusted p-values. Defaults to 0.05.}

\item{tax_levels}{Character vector. Subset of taxonomic levels: c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species", "Strain") to include in the plot.
Defaults to c("Phylum", "Genus", "Species", "Strain"). Provided levels will be reordered by the standard hierarchy and the highest level is used for colouring. If it has >30 categories,
the legend is replaced by a named vector. Only applicable with \code{aggregate_by_taxa = TRUE}.}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot. NOTE: Will be disregarded if taxonomy is provided with the option \code{aggregate_by_taxa = FALSE}.}
}
\value{
A \code{ggplot} object showing the Manhattan plot if \code{plot = TRUE}, else a \code{data.frame} with plot data.
}
\description{
This function can generate multiple Manhattan plots from a \code{strainspy_fit} object.
}
\details{
The generated plot will be more informative if taxonomic information is
provided. Two types of plots can be generated with taxonomic information.
\enumerate{
\item Set \code{aggregate_by_taxa = TRUE} to aggregate p-values across taxonomic levels
specified by \code{tax_levels}. This generates a stacked segment Manhattan plot.
In this plot, -log10 p-values adjusted using \code{method} will be plotted in
taxonomic ordering. Significance threshold will be \code{alpha}.
\item Set \code{aggregate_by_taxa = FALSE} to generate a traditional Manhattan plot for
all strains coloured by phylum. Unadjusted p-values will be plotted and
Bonferroni significance thresholds (0.05, 0.01) will be shown. Strains will
be in taxonomic ordering and a tree showing the taxonomic relationship between
strains will be plotted below the Manhattan plot.
}

If taxonomic information is not provided (\code{taxonomy = NULL}), a plot with
-log10 p-values adjusted using \code{method} will be generated. Strains will be
plotted in alphabetical order. Significance threshold will be \code{alpha}.

If Zero-Inflated Beta regression was used to model \code{object}, separate
facets will be used for zero-inflated and beta components of the model.
}
\examples{
\donttest{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
se <- filter_by_presence(se, min_nonzero = 30)
tax <- read_taxonomy(system.file("extdata", "example_taxonomy.tsv.gz", 
package = "strainspy"))
fit <- glmZiBFit(se[1:10, ], ~ Case_status + Age_at_collection, 
nthreads = 1, return_vcov = FALSE)
mh_data <- plot_manhattan(fit, taxonomy = tax, plot = FALSE)
head(mh_data)
}

}
