% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{plot_manhattan_gt}
\alias{plot_manhattan_gt}
\title{Generate a Manhattan plot from a betaGLM object, with the option to add ground truth}
\usage{
plot_manhattan_gt(
  object,
  coef = 2,
  taxonomy = NULL,
  method = "HMP",
  alpha = 0.05,
  levels = c("Phylum", "Genus", "Species", "Strain"),
  colour_level = "Phylum",
  plot = TRUE,
  ground_truth = NULL
)
}
\arguments{
\item{object}{A \code{betaGLM} object.}

\item{coef}{The number of the coefficient from which to generate the plot (default=2).}

\item{taxonomy}{An optional taxonomy file read using strainspy::read_taxonomy (default=NULL).}

\item{method}{Character. Multiple testing correction method for p-values (e.g., "holm"). Defaults to "holm".}

\item{alpha}{Numeric. Significance threshold for adjusted p-values. Defaults to 0.05.}

\item{levels}{Character vector. Taxonomic levels to include in the plot when \code{taxonomy} is provided. Defaults to c("Phylum", "Genus", "Species", "Strain").}

\item{colour_level}{Character. Taxonomic level used for coloring the plot segments. Defaults to "Phylum".}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot.}

\item{ground_truth}{A vector of ground truth contigs}
}
\value{
A \code{ggplot} object showing the Manhattan plot.
}
\description{
This function is copied from plot_manhattan()
}
