% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan_tree}
\alias{plot_manhattan_tree}
\title{Generate a \code{traditional} Manhattan plot for the strain all strains.}
\usage{
plot_manhattan_tree(object, taxonomy, coef = 2)
}
\arguments{
\item{object}{A \code{betaGLM} object.}

\item{taxonomy}{Taxonomy file read using strainspy::read_taxonomy (default=NULL).}

\item{coef}{The number of the coefficient from which to generate the plot (default=2).}
}
\value{
A \code{ggplot} object showing the Manhattan plot combined with the taxonomic tree
}
\description{
Generate a \code{traditional} Manhattan plot for the strain all strains.
}
