% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{plot_violin}
\alias{plot_violin}
\title{Generate a Violin plot for a subset of contigs for a given categorical phenotype}
\usage{
plot_violin(
  se,
  phenotype,
  contigs,
  contig_names = NULL,
  drop_ANI_zeros = FALSE,
  plot = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{phenotype}{One variable in \code{colnames(se@colData)}, except \code{Sample_File}. The selected variable must be categorical}

\item{contigs}{A vector of contigs to generate violin plots}

\item{contig_names}{A vector of names for the contigs to use in the plot, useful when the original contig names are too long. Must be the same length and order as \code{contig_names}. Default NULL.}

\item{drop_ANI_zeros}{Bool. If TRUE, ANI 0 values will not be included in the violin plot. Default FALSE. A warning will be generated to inform that ANI 0 rows are dropped.}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot.}
}
\description{
Generate a Violin plot for a subset of contigs for a given categorical phenotype
}
