% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generate a Volcano Plot from a betaGLM Object}
\usage{
plot_volcano(
  object,
  coef = 2,
  alpha = 0.05,
  method = "holm",
  palette = rev(viridis::viridis(3)),
  label = FALSE,
  contig_names = NULL,
  plot = TRUE
)
}
\arguments{
\item{object}{A strainspy_fit object containing model results.}

\item{coef}{Integer, specifying the phenotype index of the betaGLM to extract p values and effect size for plotting. Default is 2.}

\item{alpha}{Float indicating signficance level. Default is 0.05}

\item{method}{The method for p-value adjustment. For available options, see \code{?top_hits()}. Default \code{holm}.}

\item{palette}{A vector of user defined palette to colour plot. Default is the viridis palette.}

\item{label}{Logical, whether to label most signifcant points of the plot with Species name.}

\item{contig_names}{Optional character vector to replace long contig names in the plot.
Must match the length and order of \code{top_hits(object, coef = coef, alpha = alpha)$Contig_name}.
If NULL (default), shortening will be attempted automatically.}

\item{plot}{Logical, whether to return a ggplot object (default: TRUE). If FALSE, returns the processed data.}
}
\value{
A ggplot2 object if plot = TRUE, otherwise a dataframe containing the relevant statistics.
}
\description{
This function creates a volcano plot from the results of a betaGLM analysis,
highlighting the relationship between effect size (coefficients) and statistical significance (-log10 p-values).
}
\examples{
\donttest{
example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)
design <- as.formula(" ~ Case_status + Sex")
fit <- glmZiBFit(se[1:10,], design, nthreads=2)
plot_volcano(fit, coef = 2, label = TRUE)
}

}
