% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prep_for_prediction}
\alias{prep_for_prediction}
\title{Prepare a caret-compatible prediction data frame from a SummarizedExperiment}
\usage{
prep_for_prediction(
  sy,
  outcome,
  contigs,
  covariates = NULL,
  use_genome_names = TRUE
)
}
\arguments{
\item{sy}{A \link[SummarizedExperiment]{SummarizedExperiment} A
\code{SummarizedExperiment} object containing the assay data and metadata.}

\item{outcome}{Character string giving the name of the outcome variable
present in \code{colData(sy)}. This is usually the left-hand side of the
prediction model formula.}

\item{contigs}{Character vector of contigs (feature) to include as
predictors. Must be present in \code{rownames(sy)}.}

\item{covariates}{Optional character vector of additional metadata variables
to include as covariates. Default  \code{NULL} for no covariates.}

\item{use_genome_names}{Bool. Use \code{Genome_file} instead of \code{Contig_name} in
\code{colnames(output)}. \code{Genome_file} is extracted from
\code{SummarizedExperiemnt::elementMetadata(sy)}. If not available, \code{Contig_name}
will be used. Default \code{True}.}
}
\value{
A \code{data.frame} containing the specified outcome variable, optional
covariates, and selected contig features. The object is ready for direct
use in \pkg{caret} functions such as \code{caret::train()}.
}
\description{
Formats strain-level data and optional covariates into a data.frame object
ready for modeling with \pkg{caret}. The function extracts selected contigs
from the \code{sy} assay matrix, combines them with the specified \code{outcome} and
optional \code{covariates} from the metadata, and performs basic validation
and cleaning. Metadata is pulled from \code{SummarizedExperiement::colData(sy)},
and generally assumes \code{strainspy::modify_metadata()} was used upstream.
}
\details{
Character variables are automatically converted to factors.
Numeric variables are returned unchanged. For pre processing, refer to
\code{caret::train()}.
}
\examples{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- read.csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
sy <- read_sylph(example_path, example_meta)
contigs <- rownames(sy)[1:10]
df <- prep_for_prediction(
  sy = sy,
  outcome = "Case_status",
  contigs = contigs,
  covariates = c("Sex", "BMI")
)

}
