% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_strainspy_priors-class.R
\name{print,strainspy_priors-method}
\alias{print,strainspy_priors-method}
\title{Print method for strainspy_priors objects}
\usage{
\S4method{print}{strainspy_priors}(x, ...)
}
\arguments{
\item{x}{An object of class \code{strainspy_priors}}

\item{...}{Additional arguments (unused)}
}
\value{
Prints a concise summary of the object
}
\description{
Print method for strainspy_priors objects
}
