% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sylph.R
\name{read_sylph}
\alias{read_sylph}
\title{Read Sylph query or profile output file and create a SummarizedExperiment object}
\usage{
read_sylph(
  file_path,
  meta_data = NULL,
  variable = "Adjusted_ANI",
  min_identity = 95,
  clean_names = TRUE
)
}
\arguments{
\item{file_path}{Character. Path to the Sylph query or profile output file(s) in \code{tsv} format.
A single merged sylph output file processed with \code{merge_sylph_files()}, or a character vector of paths
to individual sylph output files can be provided. If multiple files are provided, they will be merged internally and saved to a temporary file.
If the internal merging fails, use \code{merge_sylph_files()} to merge the files externally and provide the merged file path.}

\item{meta_data}{data.frame. An optional tibble or data frame containing sample metadata. See Details}

\item{variable}{Character. Name of the input variable to import. Defaults to \code{Adjusted_ANI}, other options can include \code{Naive_ANI}, \code{Taxonomic_abundance} and \code{Sequence_abundance}.}

\item{min_identity}{Numeric. Minimum identity threshold for filtering ANI values. Defaults to \code{95}.}

\item{clean_names}{Logical. If \code{TRUE}, file paths will be stripped of their directory path and file extension,
leaving only the base file name. Defaults to \code{TRUE}.}
}
\value{
A SummarizedExperiment object with the following components:
\item{assays}{A matrix containing numeric features such as \code{Adjusted_ANI}, \code{Taxonomic_abundance}, \code{Median_cov}, etc.}
\item{rowData}{DataFrame containing metadata for each sequence, such as \code{Contig_name} and \code{Genome_file}.}
\item{colData}{DataFrame containing information for each sample (derived from the \code{Sample_file} field).}
}
\description{
This function reads a Sylph output file (either query or profile), parses the data, and returns a SummarizedExperiment object.
An optional metadata table can be provided and added directly to \code{colData} via an internal call to \code{modify_metadata()}.
}
\details{
If metadata is provided, it must meet the input requirements described in
\code{modify_metadata()}. If appending metadata fails due to pre processing requirements,
this function will issue a detailed warning and still return the \code{SummarizedExperiment}
object without metadata. In that case, apply necessary fixes and subsequently
call \code{modify_metadata()}.
}
\examples{

  # Read a Sylph file 
  example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
  package = "strainspy")
  sy <- read_sylph(example_path)
}
