% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_taxonomy.R
\name{read_taxonomy}
\alias{read_taxonomy}
\title{Read and process a taxonomy file in Sylph format}
\usage{
read_taxonomy(file_path)
}
\arguments{
\item{file_path}{Path to the taxonomy file.}
}
\value{
A data.table containing parsed taxonomy information.
}
\description{
This function reads a taxonomy file in Sylph format.
}
\examples{
tax <- read_taxonomy(system.file("extdata", "example_taxonomy.tsv.gz", 
package = "strainspy"))
head(tax)
}
