% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{rescale_beta}
\alias{rescale_beta}
\title{Transform ANI Data to Simulate a Treatment Effect}
\usage{
rescale_beta(x, beta = 0.98, zi = 0.1)
}
\arguments{
\item{x}{a vector of sequence identity values to transform}

\item{beta}{simulate a decreased in identiy by factor of beta (default=0.95)}

\item{zi}{simulate an increased shift in the proportion of zeros by zi (default=0.1)}
}
\value{
a vector of transformed values
}
\description{
This function applies a simple linear transformation to simulate a treatment effect.
}
