% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\name{show,betaGLM-method}
\alias{show,betaGLM-method}
\title{Show Method for betaGLM}
\usage{
\S4method{show}{betaGLM}(object)
}
\arguments{
\item{object}{An object of class `betaGLM`.}
}
\description{
Displays a summary of a `betaGLM` object, including the number of coefficients.
}
