% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_priors-class.R
\name{show,strainspy_priors-method}
\alias{show,strainspy_priors-method}
\title{Show method for strainspy_priors objects}
\usage{
\S4method{show}{strainspy_priors}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_priors}}
}
\value{
Invisibly returns the object
}
\description{
Show method for strainspy_priors objects
}
