% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_random_effects}
\alias{strip_random_effects}
\title{Remove random effects from model}
\usage{
strip_random_effects(formula)
}
\arguments{
\item{formula}{Formula object}
}
\value{
A formula without mixed components
}
\description{
Remove random effects from model
}
