% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\name{summary,betaGLM-method}
\alias{summary,betaGLM-method}
\title{Summary Method for betaGLM}
\usage{
\S4method{summary}{betaGLM}(object)
}
\arguments{
\item{object}{An object of class \code{betaGLM}.}
}
\value{
A list summarizing the \code{betaGLM} object.
}
\description{
Provides a high-level summary of a \code{betaGLM} object, including its call,
coefficients, zero-inflation coefficients, dispersion, and the availability
of FDR and residuals.
}
