% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_strainspy_fit-class.R
\name{summary,strainspy_fit-method}
\alias{summary,strainspy_fit-method}
\title{Summary Method for strainspy_fit}
\usage{
\S4method{summary}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A list summarizing the \code{strainspy_fit} object.
}
\description{
Provides a high-level summary of a \code{strainspy_fit} object, including its call,
coefficients, zero-inflation coefficients, dispersion, and the availability
of FDR and residuals.
}
