% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_hits.R
\name{top_hits}
\alias{top_hits}
\title{Provides the top results of a strainspy_fit object, sorted by adjusted p-value.}
\usage{
top_hits(fit, coef = 2, method = "holm", alpha = 0.05)
}
\arguments{
\item{fit}{A \code{strainspy_fit} object.}

\item{coef}{The coefficient to report and use for p-value adjustment. Defaults to 2.}

\item{method}{The method for p-value adjustment. Available methods are a subset of those in \code{p.adjust},
specifically those valid for dependent tests. Common choices include "holm", "hochberg", "hommel",
and "BH" (Benjamini-Hochberg) and "BY".}

\item{alpha}{Numeric. Significance threshold for adjusted p-values. Defaults to 0.05.}
}
\value{
A tibble with the top hits, sorted by adjusted p-value.
}
\description{
This function performs multiple testing adjustment on the results of a strainspy_fit object
integrating taxonomic information. Various correction methods are available,
including Harmonic Mean P-value (HMP), Bonferroni and Benjamini-Hochberg (BH).
}
\examples{
meta <- read.csv(system.file("extdata", "example_metadata.csv.gz",
 package = "strainspy"))
meta$Case_status <- factor(meta$Case_status)
se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy"), meta_data = meta)
fit <- glmZiBFit(filter_by_presence(se, min_nonzero = 30)[1:10, ], 
~ Case_status + Age_at_collection, nthreads = 1, return_vcov = FALSE)
top_hits(fit, coef = 2, alpha = 1)
}
