% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseControlFit.R
\name{caseControlFit}
\alias{caseControlFit}
\title{caseControlFit}
\usage{
caseControlFit(
  se,
  design,
  min_identity = 0.98,
  nthreads = 1,
  scale_continuous = TRUE,
  BPPARAM = NULL,
  progress = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{design}{Formula. A formula to specify the fixed and random effects, e.g., \code{Outcome ~ Value + Covariate + (1|Sample)}.}

\item{min_identity}{A numeric value specifying the minimum identity threshold to consider (default=0.98).}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{scale_continuous}{Logical. If \code{TRUE}, all numeric columns in \code{colData(se)} are z-score standardized (mean = 0, SD = 1). Defaults to \code{FALSE}.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}

\item{progress}{Logical. If \code{TRUE} (default), progress bars and progress messages are printed. Set to \code{FALSE} to silence them.}
}
\value{
A \code{strainspy_fit} object with the following components:
\item{row_data}{A DFrame with 6 slots with feature details}
\item{coefficients}{A DFrame with coefficients for each feature}
\item{std_errors}{A DFrame of standard errors for each feature}
\item{p_values}{A DFrame of p-values for each feature}
\item{residuals}{A DFrame of residual vectors for each feature}
\item{convergence}{A named logical vector indicating convergence for each feature}
\item{design}{Formula used in the call to \code{caseControlFit}}
\item{call}{Call to \code{caseControlFit}}
}
\description{
This function fits a logistic regression model similar to that published in the original Sylph paper.
It takes a \code{SummarizedExperiment} object as input, along with a user-defined formula.
}
\examples{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- read.csv(example_meta_path)
example_meta$Case_status <- factor(example_meta$Case_status)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)

design <- as.formula("Case_status ~ Value + Age_at_collection")

fit <- caseControlFit(se[1:10,], design, nthreads=2)

}
