% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseControlFit.R
\name{caseControlFit}
\alias{caseControlFit}
\title{caseControlFit}
\usage{
caseControlFit(
  se,
  design,
  min_identity = 0.98,
  nthreads = 1,
  scale_continous = TRUE,
  BPPARAM = NULL
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{min_identity}{A numeric value specifying the minimum identity threshold to consider (default=0.98).}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}

\item{formula}{Formula. A formula to specify the fixed and random effects, e.g., \code{ ~ Group + (1|Sample)}.}

\item{family}{A \code{glmmTMB} family object. Defaults to \code{glmmTMB::ordbeta()}.}
}
\value{
A list with the following components:
\item{coefficients}{A data frame with coefficients, standard errors, z-values, p-values, and FDR for each feature.}
\item{fit}{The fitted \code{glmmTMB} model object.}
\item{call}{The original function call.}
}
\description{
This function fits a logistic regression model similar to that published in the original Sylph paper.
It takes a \code{SummarizedExperiment} object as input, along with a user-defined formula.
}
\examples{
\dontrun{
library(SummarizedExperiment)
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_meta$Case_status <- factor(example_meta$Case_status)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)

design <- as.formula("Case_status ~ Value + Age_at_collection")

fit <- caseControlFit(se,  design, nthreads=4)
summary(fit)

}

}
