% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_contig_names}
\alias{clean_contig_names}
\title{Automatically clean contig names for plots}
\usage{
clean_contig_names(contig_names, max_length = 50, return_map = F)
}
\arguments{
\item{contig_names}{Vector of contig names. These will be cleaned up. This function usually works with long sylph contigs, such as those given by \code{top_hits(fit_ANI)$Contig_name}}

\item{max_length}{Numeric. Number of characters to keep (default 50)}

\item{return_map}{bool. Return data.frame of original and shortened contig names for mapping}
}
\value{
A data.frame or original and shortened names, or a vector of shortened name in the same order
}
\description{
Automatically clean contig names for plots
}
