% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_priors.R
\name{compute_eb_priors}
\alias{compute_eb_priors}
\title{Constructor for strainspy_priors using empirical Bayes estimation}
\usage{
compute_eb_priors(
  se,
  design,
  nthreads = 1L,
  BPPARAM = NULL,
  low_cutoff = 0.1,
  high_cutoff = 10,
  est_disperion_prior = FALSE
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object with count matrix and colData.}

\item{design}{A fixed-effects-only formula (e.g., ~ group + age + sex).}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}

\item{low_cutoff}{Ceiling for small SD estimates. Default 0.1}

\item{high_cutoff}{Floor for large SD estimates. Default 10}

\item{est_disperion_prior}{Estimate a prior for disperion (default = FALSE)}
}
\value{
An object of class \code{strainspy_priors}.
}
\description{
Fits fast binomial and beta models to estimate MAP prior SDs for use in
zero-inflated and ordinal beta models fitted using \code{GLMMTMB}.
}
\examples{
if (requireNamespace("fastglm", quietly = TRUE) && 
requireNamespace("betareg", quietly = TRUE)) {
  meta <- read.csv(system.file("extdata", "example_metadata.csv.gz", 
  package = "strainspy"))
  meta$Case_status <- factor(meta$Case_status)
  se <- read_sylph(system.file("extdata", "example_sylph_profile.tsv.gz", 
  package = "strainspy"), meta_data = meta)
  se <- filter_by_presence(se, min_nonzero = 30)
  pri <- compute_eb_priors(se[1:10, ], 
  ~ Case_status + Age_at_collection, nthreads = 1)
  class(pri)
}

}
