% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_priors.R
\name{compute_eb_priors}
\alias{compute_eb_priors}
\title{Constructor for strainspy_priors using empirical Bayes estimation}
\usage{
compute_eb_priors(se, design, nthreads = 1L, BPPARAM = NULL)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object with count matrix and colData.}

\item{design}{A fixed-effects-only formula (e.g., ~ group + age + sex).}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}
}
\value{
An object of class \code{strainspy_priors}.
}
\description{
Fits fast binomial and beta models to estimate MAP prior SDs for use in
zero-inflated and ordinal beta models fitted using \code{GLMMTMB}.
}
