% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_priors.R
\name{define_priors}
\alias{define_priors}
\title{Constructor for strainspy_priors using preset or user defined priors}
\usage{
define_priors(se, design, method = "preset_weak", priors_df = NULL)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object with count matrix and colData.}

\item{design}{A fixed-effects-only formula (e.g., ~ group + age + sex).}

\item{method}{Prior type ("preset_weak", "preset_strong", "manual"). To compute
empirical priors, use \code{compute_eb_priors()}. Default \code{preset_weak}.}

\item{priors_df}{NULL if \code{method == 'preset_weak' | 'preset_strong'}, else a
data.frame of priors suitable for glmmTMB model fitting. See \code{?glmmTMB::priors}
for more details. Default NULL.}
}
\value{
An object of class \code{strainspy_priors}.
}
\description{
Constructor for strainspy_priors using preset or user defined priors
}
