% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_beta.R
\name{estimate_beta_diversity}
\alias{estimate_beta_diversity}
\title{Estimate beta diversity from ANI}
\usage{
estimate_beta_diversity(
  se,
  method = "manhattan",
  distance_only = FALSE,
  ani_threshold = 95,
  phenotype = NULL,
  show_plots = FALSE,
  return_plots = FALSE
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{method}{Distance method: \code{"manhattan"} (default) or \code{"jaccard"}}

\item{distance_only}{Logical. If TRUE, stop after computing distance matrix}

\item{ani_threshold}{Numeric. Threshold for presence/absence determination if \code{method == jaccard} (default 95)}

\item{phenotype}{Any factor variable in \code{colnames(se@colData)}. Not required if \code{distance_only == FALSE}}

\item{show_plots}{Logical. If TRUE, shows the plots (default = FALSE)}

\item{return_plots}{Logical. If TRUE, returns ggplot2 objects (default = FALSE)}
}
\value{
If distance_only = TRUE, returns distance object. Otherwise a list with:
\itemize{
\item \code{distance} — distance matrix
\item \code{nmds} — NMDS object
\item \code{ordination} — NMDS coordinates with group
\item \code{permanova} — PERMANOVA results
\item \code{betadisper} — betadisper object
\item \code{plots} — list of ggplot2 objects (if return_plots = TRUE)
}
}
\description{
Estimate beta diversity from ANI
}
