% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_strainspy_priors-class.R
\name{extract_fixef_priors}
\alias{extract_fixef_priors}
\alias{extract_fixef_priors,strainspy_priors-method}
\title{Access fixef priors needed for ordinal beta regression}
\usage{
extract_fixef_priors(object)

\S4method{extract_fixef_priors}{strainspy_priors}(object)
}
\arguments{
\item{object}{A \code{strainspy_priors} object.}
}
\value{
A data frame of prior parameters for fixed effects.
}
\description{
Access fixef priors needed for ordinal beta regression
}
