% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_strainspy.R
\name{extract_strainspy}
\alias{extract_strainspy}
\title{Extract data from a StrainSpy \code{SummarizedExperiment} object in long format}
\usage{
extract_strainspy(
  se,
  variables = NULL,
  contigs = NULL,
  taxonomy = NULL,
  drop_zeros = FALSE
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{variables}{One or multiple variables in \code{colnames(se@colData)}, except \code{Sample_File}.}

\item{contigs}{A vector of contigs to generate violin plots}

\item{taxonomy}{An optional taxonomy file read using strainspy::read_taxonomy (default=NULL).}

\item{drop_zeros}{Bool. If TRUE, ANI or Abundance 0 values will not be included in the violin plot. Default FALSE.}
}
\value{
A long-format \code{data.frame} with sample metadata, contig identity columns,
ANI values, row-level annotations, and optional taxonomy columns.
}
\description{
Extract data from a StrainSpy \code{SummarizedExperiment} object in long format
}
\examples{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- read.csv(example_meta_path)
# Change variable from character to factor
example_meta$Case_status = factor(example_meta$Case_status)

example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path, example_meta)

tax_path <- system.file("extdata", "example_taxonomy.tsv.gz", 
package = "strainspy")
tax <- read_taxonomy(tax_path)

extract_strainspy(se, variables = c("Case_status", "Age_at_collection"), 
contigs = rownames(se)[1:5], taxonomy=tax)


}
