% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_presence.R
\name{filter_by_presence}
\alias{filter_by_presence}
\title{filter_by_presence}
\usage{
filter_by_presence(se, min_nonzero = 10)
}
\arguments{
\item{se}{A `SummarizedExperiment` object.}

\item{min_nonzero}{Integer. The minimum number of non-zero entries required to retain a row. Default is 10.}
}
\value{
A filtered `SummarizedExperiment` object with only the rows that meet the criteria.
}
\description{
This function filters a SummarizedExperiment object to keep only rows (features/strains)
that are present in a specified number of samples.
The same rows are filtered from all assays in the SummarizedExperiment.
}
\examples{
\dontrun{
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path)

# Filter to keep only rows with at least 10 non-zero entries in the Adjusted_ANI assay
filtered_se <- filter_by_presence(se)

# View the filtered object
filtered_se
}

}
