% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_presence.R
\name{filter_by_presence_and_rescale}
\alias{filter_by_presence_and_rescale}
\title{filter_by_presence_with_rescale_abundance}
\usage{
filter_by_presence_and_rescale(se, min_nonzero = 10, rescale_abundance = FALSE)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object.}

\item{min_nonzero}{Integer. The minimum number of non-zero entries required to retain a row. Default is 10.}

\item{rescale_abundance}{bool. Instead of containment ANI (e.g. sylph), if \code{assay(se)} contains relative abundance (e.g. \code{metaphlan}), set \code{rescale_abundance=TRUE}.
After filtering, this rescales the remaining abundances to ensure they sum to 100.}
}
\value{
A filtered \code{SummarizedExperiment} object with only the rows that meet the criteria.
}
\description{
This function is copied from filter_by_presence(). Additionally, it rescales abundances such that
each sample sums back to 100. Since this behaviour is not useful for association testing, this function is not exported.
}
\examples{
\donttest{
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path)

# Filter to keep only rows with at least 10 non-zero entries in the assay
filtered_se <- filter_by_presence(se)

# View the filtered object
filtered_se
}

}
