% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_fit-class.R
\name{getContigNames}
\alias{getContigNames}
\alias{getContigNames,strainspy_fit-method}
\title{Get Contig names in a fit object}
\usage{
getContigNames(object)

\S4method{getContigNames}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A vector of contig names
}
\description{
Gets the ordered list of fitted contig names in a \code{strainspy_fit}
}
