% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\name{getFDR}
\alias{getFDR}
\alias{getFDR,betaGLM-method}
\title{Access False Discovery Rates (FDR)}
\usage{
getFDR(object)

\S4method{getFDR}{betaGLM}(object)
}
\arguments{
\item{object}{An object of class \code{betaGLM}.}
}
\value{
A \code{DataFrame} or \code{NULL} containing FDR values.
}
\description{
Returns the false discovery rates (FDR) from a \code{betaGLM} object, if available.
}
