% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\name{getFDR}
\alias{getFDR}
\title{Access False Discovery Rates (FDR)}
\usage{
getFDR(object)
}
\arguments{
\item{object}{An object of class `betaGLM`.}
}
\value{
A `DataFrame` or `NULL` containing FDR values.
}
\description{
Returns the false discovery rates (FDR) from a `betaGLM` object, if available.
}
