% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_fit-class.R
\name{getFDR}
\alias{getFDR}
\alias{getFDR,strainspy_fit-method}
\title{Access False Discovery Rates (FDR)}
\usage{
getFDR(object)

\S4method{getFDR}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A \code{DataFrame} or \code{NULL} containing FDR values.
}
\description{
Returns the false discovery rates (FDR) from a \code{strainspy_fit} object, if available.
}
