% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_fit-class.R
\name{getGenomes}
\alias{getGenomes}
\alias{getGenomes,strainspy_fit-method}
\title{Get Genomes in a fit object}
\usage{
getGenomes(object)

\S4method{getGenomes}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A vector of fitted genomes
}
\description{
Gets the ordered list of fitted genomes in a \code{strainspy_fit}
}
