% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_fit-class.R
\name{getResiduals}
\alias{getResiduals}
\alias{getResiduals,strainspy_fit-method}
\title{Access Residuals}
\usage{
getResiduals(object)

\S4method{getResiduals}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A \code{DataFrame} containing residuals.
}
\description{
Returns the residuals from a \code{strainspy_fit} object.
}
