% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\name{getResiduals}
\alias{getResiduals}
\title{Access Residuals}
\usage{
getResiduals(object)
}
\arguments{
\item{object}{An object of class `betaGLM`.}
}
\value{
A `DataFrame` containing residuals.
}
\description{
Returns the residuals from a `betaGLM` object.
}
