% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_fit-class.R
\name{getSamples}
\alias{getSamples}
\alias{getSamples,strainspy_fit-method}
\title{Get Sample names in a fit object}
\usage{
getSamples(object)

\S4method{getSamples}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A vector of fitted sample names
}
\description{
Gets the ordered list of samples used in \code{strainspy_fit}
}
