% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaGLM-class.R
\name{getZICoefficients}
\alias{getZICoefficients}
\alias{getZICoefficients,betaGLM-method}
\title{Access Zero-Inflation Coefficients}
\usage{
getZICoefficients(object)

\S4method{getZICoefficients}{betaGLM}(object)
}
\arguments{
\item{object}{An object of class \code{betaGLM}.}
}
\value{
A \code{DataFrame} containing zero-inflation coefficients.
}
\description{
Returns the zero-inflation coefficients from a \code{betaGLM} object.
}
