% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_fit-class.R
\name{getZICoefficients}
\alias{getZICoefficients}
\alias{getZICoefficients,strainspy_fit-method}
\title{Access Zero-Inflation Coefficients}
\usage{
getZICoefficients(object)

\S4method{getZICoefficients}{strainspy_fit}(object)
}
\arguments{
\item{object}{An object of class \code{strainspy_fit}.}
}
\value{
A \code{DataFrame} containing zero-inflation coefficients.
}
\description{
Returns the zero-inflation coefficients from a \code{strainspy_fit} object.
}
