% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmQBFit.R
\name{glmQBFit}
\alias{glmQBFit}
\title{glmQB}
\usage{
glmQBFit(se, design, nthreads = 1L, scale_continous = TRUE, BPPARAM = NULL)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing
the assay data and metadata.}

\item{design}{Formula. A formula to specify fixed effects, e.g.,
\code{ ~ Age + BMI}.}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use.
Defaults to one 1.}

\item{scale_continous}{Logical. If \code{TRUE}, all numeric columns in
\code{colData(se)} are z-score standardized (mean = 0, SD = 1). Default \code{FALSE}.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the
function will configure an appropriate backend automatically.}
}
\value{
A \code{strainspy_fit} object with the following components:
\item{row_data}{A DFrame with 6 slots with feature details}
\item{coefficients}{A DFrame with coefficients for each feature}
\item{std_errors}{A DFrame of standard errors for each feature}
\item{p_values}{A DFrame of p-values for each feature}
\item{residuals}{A DFrame of residual vectors for each feature}
\item{convergence}{A named logical vector indicating convergence for each feature}
\item{design}{Formula used in the call to \code{glmQBFit}}
\item{call}{Call to \code{glmQBFit}}
}
\description{
This function fits a quasibinomial regression model using \code{stats::glm()}.

It takes a \code{SummarizedExperiment} object as input, along with a user-defined
formula containing only fixed effect parameters, and fits a quasibinomial
regression model on the assay data.
}
\examples{
\dontrun{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)

design <- as.formula(" ~ Case_status + Age_at_collection")

fit <- glmQB(se,  design, nthreads=parallel::detectCores())
top_hits(fit, alpha=0.5)
plot_manhattan(fit)

}

}
