% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmZiBFit.R
\name{glmZiBFit}
\alias{glmZiBFit}
\title{glmZiBFit}
\usage{
glmZiBFit(se, design, nthreads = 1, scale_continous = TRUE, BPPARAM = NULL)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}

\item{formula}{Formula. A formula to specify the fixed and random effects, e.g., \code{ ~ Group + (1|Sample)}.}
}
\value{
A list with the following components:
\item{coefficients}{A data frame with coefficients, standard errors, z-values, p-values, and FDR for each feature.}
\item{fit}{The fitted \code{glmmTMB} model object.}
\item{call}{The original function call.}
}
\description{
This function fits a zero-inflated beta regression model using the \code{glmmTMB} package.
It takes a \code{SummarizedExperiment} object as input, along with a user-defined formula,
and fits a zero-inflated beta regression model on the assay data.
}
\examples{
\dontrun{
library(SummarizedExperiment)
library(strainspy)
library(glmmTMB)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)

design <- as.formula(" ~ Case_status + Age_at_collection")

fit <- glmZiBFit(se,  design, nthreads=4)
summary(fit)

}

}
