% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmZiBFit.R
\name{glmZiBFit}
\alias{glmZiBFit}
\title{glmZiBFit}
\usage{
glmZiBFit(
  se,
  design,
  nthreads = 1,
  scale_continous = TRUE,
  MAP_prior = "preset_weak",
  BPPARAM = NULL,
  method = "glmmTMB"
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{design}{Formula. A formula to specify the fixed and random effects, e.g., \code{ ~ Group + (1|Sample)}.}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults
to one 1.}

\item{scale_continous}{Logical. If \code{TRUE}, all numeric columns in \code{colData(se)} are z-score standardized (mean = 0, SD = 1). Defaults to \code{FALSE}.}

\item{MAP_prior}{One of \code{strainspy_priors} object, character string (\verb{"preset_weak", "preset_strong"}),
a \code{data.frame} of priors, or \code{NULL}. Default \code{"preset_weak"}. See Details.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function
will configure an appropriate backend automatically.}

\item{method}{Character. The method to use for fitting the model. Either 'glmmTMB' (default) or 'gamlss'.}
}
\value{
A \code{strainspy_fit} object with the following components:
\item{row_data}{A DFrame with 6 slots with feature details}
\item{coefficients}{A DFrame with coefficients for each feature}
\item{std_errors}{A DFrame of standard errors for each feature}
\item{p_values}{A DFrame of p-values for each feature}
\item{zi_coefficients}{A DFrame of zero inflated coefficients for each feature}
\item{zi_std_errors}{A DFrame of zero inflated standard errors for each feature}
\item{zi_p_values}{A DFrame of zero inflated p-values for each feature}
\item{residuals}{A DFrame of residual vectors for each feature}
\item{convergence}{A named logical vector indicating convergence for each feature}
\item{design}{Formula used in the call to \code{glmZiBFit}}
\item{call}{Call to \code{glmZiBFit}}
}
\description{
This function fits a zero-inflated beta regression model using either \code{glmmTMB} or \code{gamlss} packages.
It takes a \code{SummarizedExperiment} object as input, along with a user-defined formula,
and fits a zero-inflated beta regression model on the assay data.
}
\examples{
\dontrun{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
se <- filter_by_presence(se)

design <- as.formula(" ~ Case_status")

fit <- glmZiBFit(se, design, nthreads=parallel::detectCores())
top_hits(fit, alpha=0.5)
plot_manhattan(fit)

}

}
