% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadjust.R
\name{hadjust}
\alias{hadjust}
\title{Performs hierarchical taxonomy aware adjustment of p-values.}
\usage{
hadjust(object, coef = 2, method = "HMP", taxonomy = NULL, index_range = FALSE)
}
\arguments{
\item{object}{A \code{betaGLM} object.}

\item{coef}{The coefficient to use for p-value adjustment. Defaults to 2.}

\item{method}{The method for p-value adjustment. Options include "bonferroni" and Harmonic Mean P-value (HMP).}

\item{taxonomy}{A taxonomy data.table object. If provided, the p-values will be adjusted at each taxonomic level.}

\item{index_range}{description}
}
\value{
A tibble with original and adjusted p-values.
}
\description{
This function performs multiple testing adjustment on the results of a betaGLM object
integrating taxonomic information. Various correction methods are available,
including Harmonic Mean P-value (HMP), Bonferroni and Benjamini-Hochberg (BH).
}
