% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash.R
\name{merge_sourmash_files}
\alias{merge_sourmash_files}
\title{merge_sourmash_files}
\usage{
merge_sourmash_files(
  sourmash_files,
  manifest_file = NULL,
  output,
  strip_unusable = T
)
}
\arguments{
\item{sourmash_files}{A character vector of file paths to merge.}

\item{manifest_file}{If one or several indexed databases were used for searching or gathering, required genome column can be missing.
The path to the output of \verb{sourmash sig manifest} can be provided to try and add back the data automatically (Default = NULL).
If provided, columns \code{md5}, \code{name} and \code{filename} are required.  This function uses \code{name} and \code{md5} to match the manifest data with sourmash outputs.
In the final output, \code{filename} will be used to identify each hit. If the \code{filename} is a path, the basename without extensions will be used instead.}

\item{output}{Save path for the output file. Interally, data.table::fwrite is used, provide the extension \code{gz} to save a compressed file.}

\item{strip_unusable}{Remove columns containing data not useful for further analysis. Default T.}
}
\description{
Merge the outputs of \verb{sourmash search} or \verb{sourmash gather} into a single file.
The output of this function is suitable to use in read_sourmash()
}
\details{
Refer to \href{https://sourmash.readthedocs.io/en/latest/classifying-signatures.html#searching-for-similar-samples-with-search}{sourmash documentation} for more information.
}
