% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sylph.R
\name{merge_sylph_files}
\alias{merge_sylph_files}
\title{merge_sylph_files}
\usage{
merge_sylph_files(sylph_files, output)
}
\arguments{
\item{sylph_files}{A character vector of sylph output file paths to merge.}

\item{output}{Save path for the output file. Internally, data.table::fwrite is used, provide the extension \code{gz} to save a compressed file.}
}
\value{
Invisibly returns \code{NULL}. The merged sylph table is written to \code{output}.
}
\description{
Merge the outputs of \verb{sylph query} or \verb{sylph profile} into a single file.
The output of this function is suitable to use in read_sylph()
}
