% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{modify_metadata}
\alias{modify_metadata}
\title{Add or modify metadata in an \code{SummarizedExperiment} object}
\usage{
modify_metadata(se, meta_data, replace = F)
}
\arguments{
\item{se}{SummarizedExperiment object generated from \code{strainspy} read functions.}

\item{meta_data}{data.frame. A tibble or data frame containing sample metadata.}

\item{replace}{bool. If T, the meta data in se (i.e., \code{se@coldata}) will be replaced with the provided meta_data file. Default F.
The \strong{first column must contain sample names} that match exactly with the \code{colnames(se)}.}
}
\value{
SummmarizedExperiment object \code{se}, updated with the new meta data.
}
\description{
This function adds or modifies metadata in \code{SummarizedExperiment} object read using functions such as \code{read_sylph()} and \code{read_metaphlan()}.
}
