% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_metadata.R
\name{modify_metadata}
\alias{modify_metadata}
\title{Add or modify metadata in a \code{SummarizedExperiment} object}
\usage{
modify_metadata(se, meta_data, replace = FALSE, convert = TRUE)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{meta_data}{A data.frame with sample metadata. \strong{The first column must contain all samples returned by} \code{colnames(se)}.}

\item{replace}{Logical. If TRUE, replaces existing colData. Default FALSE.}

\item{convert}{Logical. If TRUE, attempts to auto-convert types as suited for
model fitting (int to numeric, char to factor). Disable if meta_data is fully
configured for model fitting. Default TRUE. See Details.}
}
\value{
Updated SummarizedExperiment object.
}
\description{
This function adds or modifies metadata in a \code{SummarizedExperiment} object.
}
\details{
If \code{convert = TRUE}, character columns with less than 5 unique (non-NA) values will be automatically
converted to factors, with levels ordered alphabetically. For meaningful association testing,
however, factor levels should reflect the intended reference level. It is recommended to
manually convert such variables to factors with explicitly set levels \emph{before} calling \code{modify_metadata()}.
}
\examples{
\dontrun{
  # Add or modify metadata subsequently
  example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
  example_meta <- readr::read_csv(example_meta_path)
  example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
  se <- read_sylph(example_path)
  se <- modify_metadata(se, example_meta)
}

}
