% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ani_distribution.R
\name{plot_ani_dist}
\alias{plot_ani_dist}
\title{Generate a violin or box plot for a subset of contigs for a categorical phenotype.}
\usage{
plot_ani_dist(
  se,
  phenotype,
  contigs,
  contig_names = NULL,
  drop_zeros = FALSE,
  show_points = FALSE,
  plot = TRUE,
  plot_type = "violin"
)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{phenotype}{One variable in \code{colnames(se@colData)}, except \code{Sample_File}. The selected variable must be categorical}

\item{contigs}{A vector of contigs to generate violin plots}

\item{contig_names}{Optional character vector to replace long contig names in the plot.
Must match the length and order of \code{contigs}. If NULL (default), shortening will be attempted automatically.}

\item{drop_zeros}{Bool. If TRUE, ANI or Abundance 0 values will not be included in the violin plot. Default FALSE.}

\item{show_points}{Bool. If TRUE, points will be overlaid with jitter using ggforce::geom_sina(). Default FALSE.}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot.}

\item{plot_type}{Specify whether \code{violin} or \code{box} plot is required. Default \code{violin}.}
}
\description{
Generate a violin or box plot for a subset of contigs for a categorical phenotype.
}
\examples{
\dontrun{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
# Change variable from character to factor
example_meta$Case_status = factor(example_meta$Case_status)

example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)

plot_ani_dist(se, phenotype = "Case_status", contigs = rownames(se)[1:5])


}

}
