% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ani_distribution.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Generate a histogram for a single contig for a categorical phenotype.}
\usage{
plot_histogram(se, phenotype, contig, drop_zeros = F)
}
\arguments{
\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{phenotype}{One variable in \code{colnames(se@colData)}, except \code{Sample_File}. The selected variable must be categorical}

\item{contig}{Contig name to visualise}

\item{drop_zeros}{Bool. If TRUE, ANI or Abundance 0 values will not be included in the violin plot. Default FALSE}
}
\description{
Generate a histogram for a single contig for a categorical phenotype.
}
\examples{
\dontrun{
library(strainspy)

example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
# Change variable from character to factor
example_meta$Case_status = factor(example_meta$Case_status)

example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)

plot_histogram(se, phenotype = "Case_status", contigs = rownames(se)[1])

}

}
