% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Generate a strain Manhattan plot from a betaGLM object}
\usage{
plot_manhattan(
  object,
  coef = 2,
  taxonomy = NULL,
  method = "holm",
  alpha = 0.05,
  tax_levels = c("Phylum", "Genus", "Species", "Strain"),
  plot = TRUE
)
}
\arguments{
\item{object}{A \code{betaGLM} object.}

\item{coef}{The number of the coefficient from which to generate the plot (default=2).}

\item{taxonomy}{An optional taxonomy file read using strainspy::read_taxonomy (default=NULL).}

\item{method}{Character. Multiple testing correction method for p-values (e.g., "holm"). Defaults to "holm".}

\item{alpha}{Numeric. Significance threshold for adjusted p-values. Defaults to 0.05.}

\item{tax_levels}{Character vector. Subset of taxonomic levels: c("Domain", "Phylum", "Class", "Order", "Family", "Genus", "Species", "Strain") to include in the plot.
Defaults to c("Phylum", "Genus", "Species", "Strain"). Provided levels will be reordered by the standard hierarchy and the highest level is used for colouring. If it has >30 categories,
the legend is replaced by a named vector.}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot.}
}
\value{
A \code{ggplot} object showing the Manhattan plot.
}
\description{
This function creates a Manhattan plot from a \code{betaGLM} object, with separate facets
for the Zero-Inflation and Beta distribution parts of the model.
}
