% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhatten.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Generate a strain Manhattan plot from a betaGLM object}
\usage{
plot_manhattan(
  object,
  coef = 2,
  taxonomy = NULL,
  method = "HMP",
  alpha = 0.05,
  levels = c("Phylum", "Genus", "Species", "Strain"),
  colour_level = "Phylum",
  plot = TRUE
)
}
\arguments{
\item{object}{A `betaGLM` object.}

\item{coef}{The number of the coefficient from which to generate the plot (default=2).}

\item{plot}{If set to false, the function will return a tibble with data used to generate the plot.}
}
\value{
A `ggplot` object showing the Manhattan plot.
}
\description{
This function creates a Manhattan plot from a `betaGLM` object, with separate facets
for the Zero-Inflation and Beta distribution parts of the model.
}
