% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot MDS of ANI data}
\usage{
plot_mds(
  se,
  coef = 2,
  taxonomy = NULL,
  palette = c("#DAA520", "#3CB371", "#483D8B"),
  plot = TRUE
)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object containing ANI data.}

\item{coef}{Integer. Index of the column in \code{colData(se)} to use as the grouping or color variable (default is 2).}

\item{taxonomy}{Optional taxonomy table to use in distance calculation.}

\item{palette}{A vector of color hex codes for numeric phenotype gradients. Defaults to viridis palette.}

\item{plot}{Logical. If \code{TRUE}, returns a \code{ggplot} object. If \code{FALSE}, returns the underlying MDS coordinates as a tibble.}
}
\value{
A \code{ggplot} object showing the MDS plot, or a tibble of MDS coordinates if \code{plot = FALSE}.
}
\description{
Performs classical multidimensional scaling (MDS) on a \code{SummarizedExperiment} object
and plots the result, optionally colored by a phenotype.
}
\examples{
\dontrun{
example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
plot_mds(se, coef = 2)
}

}
