% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca_biplot.R
\name{plot_pca_biplot}
\alias{plot_pca_biplot}
\title{plot_pca_biplot}
\usage{
plot_pca_biplot(se, phenotype, plot = TRUE)
}
\arguments{
\item{se}{A \code{SummarizedExperiment} object containing expression data.}

\item{phenotype}{A character string specifying the name of a phenotype or an integer specifying the column number in colData(se)}

\item{plot}{A logical value indicating whether to return the PCA plot (\code{TRUE}) or PCA data (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
If \code{plot = TRUE}, returns a \code{ggplot2} PCA plot. If \code{plot = FALSE}, returns a tibble with PCA results.
}
\description{
Perform PCA and Generate a Biplot
}
\details{
This function performs Principal Component Analysis (PCA) on a SummarizedExperiment object
and generates a PCA biplot colored by a specified phenotype.
}
\examples{
\donttest{
example_meta_path <- system.file("extdata", "example_metadata.csv.gz", 
package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", 
package = "strainspy")
se <- read_sylph(example_path, example_meta)
plot_pca_biplot(se, phenotype = 3)
}

}
