% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy_priors-class.R
\name{plot_prior_bootstrap}
\alias{plot_prior_bootstrap}
\alias{plot_prior_bootstrap,strainspy_priors-method}
\title{Plot bootstrap prior SDs (fixef + fixef_zi) for a coefficient}
\usage{
plot_prior_bootstrap(object, prior)

\S4method{plot_prior_bootstrap}{strainspy_priors}(object, prior)
}
\arguments{
\item{object}{A \code{strainspy_priors} object.}

\item{prior}{Character. Coefficient name (must exist in both prior matrices).}
}
\value{
A patchwork side-by-side ggplot object.
}
\description{
Plot bootstrap prior SDs (fixef + fixef_zi) for a coefficient
}
