% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generate a Volcano Plot from a betaGLM Object}
\usage{
plot_volcano(
  fit,
  coef = 2,
  alpha = 0.05,
  palette = c("#DAA520", "#3CB371", "#483D8B"),
  label = FALSE,
  regex = "\\\\b[A-Z][a-z]+ [a-z]+(?: [a-z]+)?(?: [A-Z]{2,}[^\\\\s,]*)*",
  plot = TRUE
)
}
\arguments{
\item{fit}{A \code{betaGLM} object containing model results.}

\item{coef}{Integer, specifying the phenotype index of the betaGLM to extract p values and effect size for plotting. Default is 2.}

\item{alpha}{Float indicating signficance level}

\item{palette}{A vector of user defined palette to colour plot. Default is the viridis palette.}

\item{label}{Logical, whether to label most signifcant points of the plot with Species name.}

\item{regex}{String, a regular expression which can be used to format the labels}

\item{plot}{Logical, whether to return a ggplot object (default: TRUE). If FALSE, returns the processed data.}
}
\value{
A ggplot2 object if \code{plot = TRUE}, otherwise a dataframe containing the relevant statistics.
}
\description{
This function creates a volcano plot from the results of a betaGLM analysis,
highlighting the relationship between effect size (coefficients) and statistical significance (-log10 p-values).
}
\examples{
\dontrun{
example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
example_meta <- readr::read_csv(example_meta_path)
example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
se <- read_sylph(example_path, example_meta)
design <- as.formula(" ~ Case_status + Sex")
fit <- glmZiBFit(se[1:100,], design, nthreads=1, method='gamlss')
plot_volcano(fit, coef = 2)
}

}
