% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metaphlan.R
\name{read_metaphlan}
\alias{read_metaphlan}
\title{read_metaphlan}
\usage{
read_metaphlan(
  file_path,
  meta_data = NULL,
  variable = "relative_abundance",
  clean_names = TRUE
)
}
\arguments{
\item{file_path}{Character. Path to the merged metaphlan output file.}

\item{meta_data}{data.frame. An optional tibble or data frame containing sample metadata. See Details}

\item{variable}{Character. Name of the input variable to import. Defaults to \code{relative_abundance}.}

\item{clean_names}{Logical. If \code{TRUE}, file paths will be stripped of their directory path and file extension,
leaving only the base file name. Defaults to \code{TRUE}.}
}
\value{
A SummarizedExperiment object with the following components:
\item{assays}{A matrix containing numeric features such as \code{relative_abundance}}
\item{rowData}{DataFrame containing metadata for each sequence, such as \code{Contig_name} and \code{Genome_file}.}
\item{colData}{DataFrame containing information for each sample (derived from the \code{Sample_file} field).}
}
\description{
This function reads a merged \code{MetaPhlAn} profile output and create a SummarizedExperiment object.
An optional metadata table can be provided and added directly to \code{colData} via an internal call to \code{modify_metadata()}.
}
\note{
Since \code{MetaPhlAn} profiles are usually generated per sample, first use \code{merge_metaphlan_files()} to merge and generate the input for this function.

#' If metadata is provided, it must meet the input requirements described in
\code{modify_metadata()}. If appending metadata fails due to pre processing requirements,
this function will issue a detailed warning and still return the \code{SummarizedExperiment}
object without metadata. In that case, apply necessary fixes and subsequently
call \code{modify_metadata()}.
}
\examples{
\dontrun{
  # Read a merged metaphlan profile file into a SummarizedExperiment object
  example_path <- system.file("extdata", "metaphlan_merged.tsv.gz", package = "strainspy")
  mp <- read_metaphlan(example_path)
  # View the SummarizedExperiment
  mp
  # View the assays (numerical matrix)
  assay(mp)
  # View the rowData (metadata for contigs)
  rowData(mp)
  # View the colData (metadata for samples)
  colData(mp)
}
}
