% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sourmash.R
\name{read_sourmash}
\alias{read_sourmash}
\title{read_sourmash}
\usage{
read_sourmash(
  file_path,
  meta_data = NULL,
  variable = "match_containment_ani",
  clean_names = TRUE
)
}
\arguments{
\item{file_path}{Character. Path to the sourmash gather file (CSV format).}

\item{meta_data}{data.frame. A tibble or data frame containing sample metadata. Metadata requirements are identical to \code{read_sylph()}.}

\item{variable}{Character. Name of the input variable to import. Defaults to \code{match_containment_ani}.}

\item{clean_names}{Logical. If \code{TRUE}, file paths will be stripped of their directory path and file extension,
leaving only the base file name. Defaults to \code{TRUE}.}
}
\value{
A SummarizedExperiment object with the following components:
\item{assays}{A matrix containing numeric features such as \code{Adjusted_ANI}, \code{Taxonomic_abundance}, \code{Median_cov}, etc.}
\item{rowData}{DataFrame containing metadata for each sequence, such as \code{Contig_name} and \code{Genome_file}.}
\item{colData}{DataFrame containing information for each sample (derived from the \code{Sample_file} field).}
}
\description{
Read a \verb{sourmash gather} or \verb{sourmash search} -like output CSV file and create a SummarizedExperiment object.
Optionally, metadata can be provided, which will be loaded into the \code{colData} of the SummarizedExperiment object.
}
\note{
Some of sourmash subcommands like \code{gather} are designed to take exactly one query file.
To merge and save a CSV file suitable to use with this function, use \code{merge_sourmash_files()}
}
\examples{
\dontrun{
  # Read a merged sourmash file into a SummarizedExperiment object
  example_path <- system.file("extdata", "example_sourmash_merged.csv.gz", package = "strainspy")
  sm <- read_sourmash(example_path)
  # View the SummarizedExperiment
  sm
  # View the assays (numerical matrix)
  assay(sm)
  # View the rowData (metadata for contigs)
  rowData(sm)
  # View the colData (metadata for samples)
  colData(sm)
}
}
