% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sylph.R
\name{read_sylph}
\alias{read_sylph}
\title{Read Sylph query or profile output file and create a SummarizedExperiment object}
\usage{
read_sylph(
  file_path,
  meta_data = NULL,
  variable = "Adjusted_ANI",
  clean_names = TRUE
)
}
\arguments{
\item{file_path}{Character. Path to the Sylph query or profile output file (tab-separated format).}

\item{meta_data}{data.frame. A tibble or data frame containing sample metadata.
The \strong{first column must contain sample names} that match exactly with the sample names in the Sylph output.}

\item{variable}{Character. Name of the input variable to import, either \code{Adjusted_ANI} or  \code{Naive_ANI}. Defaults to \code{Adjusted_ANI}.}

\item{clean_names}{Logical. If \code{TRUE}, file paths will be stripped of their directory path and file extension,
leaving only the base file name. Defaults to \code{TRUE}.}
}
\value{
A SummarizedExperiment object with the following components:
\item{assays}{A matrix containing numeric features such as \code{Adjusted_ANI}, \code{Taxonomic_abundance}, \code{Median_cov}, etc.}
\item{rowData}{DataFrame containing metadata for each sequence, such as \code{Contig_name} and \code{Genome_file}.}
\item{colData}{DataFrame containing information for each sample (derived from the \code{Sample_file} field).}
}
\description{
This function reads a Sylph output file (either query or profile), parses the data, and returns a SummarizedExperiment object.
Optionally, metadata can be provided, which will be loaded into the \code{colData} of the SummarizedExperiment object.
}
\examples{
\dontrun{
  # Read a Sylph file (query or profile) into a SummarizedExperiment object
  example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
  se <- read_sylph(example_path)
  # View the SummarizedExperiment
  se
  # View the assays (numerical matrix)
  assay(se)
  # View the rowData (metadata for contigs)
  rowData(se)
  # View the colData (metadata for samples)
  colData(se)
}
\dontrun{
  # Read a Sylph file (query or profile) with associated metadata into a SummarizedExperiment object
  example_meta_path <- system.file("extdata", "example_metadata.csv.gz", package = "strainspy")
  example_meta <- readr::read_csv(example_meta_path)
  example_path <- system.file("extdata", "example_sylph_profile.tsv.gz", package = "strainspy")
  se <- read_sylph(example_path, example_meta)
}
}
