% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_beta.R
\name{rescale_beta}
\alias{rescale_beta}
\title{Transform ANI Data to Simulate a Treatment Effect}
\usage{
rescale_beta(x, beta = 0.95, zi = 0.1)
}
\arguments{
\item{x}{a vector of sequence identity values to transform}

\item{beta}{simulate a decreased in identiy by factor of beta (default=0.95)}

\item{zi}{simulate an increased shift in the proportion of zeros by zi (default=0.1)}
}
\value{
a vector of transformed values
}
\description{
This function applies a simple linear transformation to simulate a treatment effect.
}
\examples{
N_group <- 1000

shape1 <- 2
shape2 <- 10
base_zero <- 0.5

sim <- rbeta(N_group, shape1, shape2)
sim <- sim*rbinom(length(sim), 1, base_zero)


# Simulate a treatment effect.

beta_difference <- runif(1, 0.5, 0.9)
treat_zero_effect <- 0.2

treat <- rescale_beta(sim, beta_difference, treat_zero_effect)

df <- data.frame(cat=rep(c('a','b'), each=N_group),
                 vals=c(sim,treat$rescaled))

m <- glmmTMB::glmmTMB(vals ~ cat, data=df,
                      family=glmmTMB::beta_family(link="logit"),
                      ziformula = ~cat)

coefs <- glmmTMB::fixef(m)$cond
zi_coefs <- glmmTMB::fixef(m)$zi


}
