% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy-package.R
\docType{package}
\name{strainspy-package}
\alias{strainspy}
\alias{strainspy-package}
\title{strainspy: Strain Level Microbiome Wide Association Studies}
\description{
Strainspy fits statistical models to detect microbial strains associated with a phenotype. For small sample sizes, models can be fitted using empirical Bayes maximum a posteriori (MAP) priors to improve detection power. Supports various experimental design scenarios, including repeated measures.
}
\keyword{internal}
