% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainspy-package.R
\docType{package}
\name{strainspy-package}
\alias{strainspy}
\alias{strainspy-package}
\title{strainspy: Strain Level Microbiome Wide Association Studies}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
