% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_strainspy_fit-class.R
\docType{class}
\name{strainspy_fit-class}
\alias{strainspy_fit-class}
\title{strainspy_fit: A Class for storing fit data from all strainspy models}
\description{
The \code{strainspy_fit} class is designed to store the results of strainspy
}
\section{Slots}{

\describe{
\item{\code{row_data}}{A \code{DataFrame} containing row data.}

\item{\code{priors}}{A \code{strainspy_priors} object or \code{NULL} for REML.}

\item{\code{coefficients}}{A \code{DataFrame} containing model coefficients.}

\item{\code{std_errors}}{A \code{DataFrame} containing standard errors of the coefficients.}

\item{\code{p_values}}{A \code{DataFrame} containing p-values for the coefficients.}

\item{\code{zi_coefficients}}{A \code{DataFrame} containing zero-inflation coefficients.}

\item{\code{zi_std_errors}}{A \code{DataFrame} containing zero-inflation standard errors.}

\item{\code{zi_p_values}}{A \code{DataFrame} containing zero-inflation p-values.}

\item{\code{residuals}}{A \code{DataFrame} containing residuals from the model.}

\item{\code{convergence}}{A logical indicating whether model fitting converged.}

\item{\code{design}}{A matrix representing the design matrix of the model.}

\item{\code{assay}}{A matrix containing the assay data used for fitting the model.}

\item{\code{call}}{The matched call of the model.}
}}

