% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{update_fit}
\alias{update_fit}
\title{Update the model fit by calling the specified fit function only a subset of features. Same design will be fitted.}
\usage{
update_fit(
  fit,
  update_idx,
  se,
  scale_continuous = TRUE,
  min_identity = 0.98,
  method = "glmmTMB",
  family = glmmTMB::ordbeta(),
  nthreads = 1,
  BPPARAM = NULL
)
}
\arguments{
\item{fit}{A \code{strainspy_fit} object}

\item{update_idx}{Vector of indices that need to be updated}

\item{se}{SummarizedExperiment. A \code{SummarizedExperiment} object containing the assay data and metadata.}

\item{scale_continuous}{Binary specifying whether to rescale numeric values. Default T}

\item{min_identity}{Only for \code{caseControlFit()}. A numeric value specifying the minimum identity threshold to consider (default=0.98).}

\item{method}{Character. The method to use for fitting the model. Either 'glmmTMB' (default) or 'gamlss'. Only applicable for \code{glmZiBFit()}.}

\item{family}{A \code{glmmTMB} family object. Defaults to \code{glmmTMB::ordbeta()}. Only applicable for \code{glmZiBFit()}}

\item{nthreads}{An integer specifying the number of (CPUs or workers) to use. Defaults to 1.}

\item{BPPARAM}{Optional \code{BiocParallelParam} object. If not provided, the function will configure an appropriate backend automatically.}
}
\value{
Updated version of the fit object
}
\description{
Update the model fit by calling the specified fit function only a subset of features. Same design will be fitted.
}
